move merging logic to postprocessor
