+++ thirdeye-spi/src/main/java/ai/startree/thirdeye/spi/detection/postprocessing/PostProcessingContext.java
@@ -15,6 +15,7 @@ package ai.startree.thirdeye.spi.detection.postprocessing;
 
 import ai.startree.thirdeye.spi.datalayer.bao.DatasetConfigManager;
 import ai.startree.thirdeye.spi.datalayer.bao.MergedAnomalyResultManager;
+import ai.startree.thirdeye.spi.datalayer.dto.EnumerationItemDTO;
 import ai.startree.thirdeye.spi.datasource.loader.MinMaxTimeLoader;
 import ai.startree.thirdeye.spi.detection.DetectionPipelineUsage;
 
@@ -25,18 +26,21 @@ public class PostProcessingContext {
   private final MergedAnomalyResultManager mergedAnomalyResultManager;
   private final Long alertId;
   private final DetectionPipelineUsage usage;
+  private final EnumerationItemDTO enumerationItemDTO;
 
   public PostProcessingContext(
       final DatasetConfigManager datasetConfigManager,
       final MinMaxTimeLoader minMaxTimeLoader,
       final MergedAnomalyResultManager mergedAnomalyResultManager,
       final Long alertId,
-      final DetectionPipelineUsage usage) {
+      final DetectionPipelineUsage usage,
+      final EnumerationItemDTO enumerationItemDTO) {
     this.datasetConfigManager = datasetConfigManager;
     this.minMaxTimeLoader = minMaxTimeLoader;
     this.mergedAnomalyResultManager = mergedAnomalyResultManager;
     this.alertId = alertId;
     this.usage = usage;
+    this.enumerationItemDTO = enumerationItemDTO;
   }
 
   public DatasetConfigManager getDatasetConfigManager() {
@@ -57,5 +61,9 @@ public class PostProcessingContext {
 
   public DetectionPipelineUsage getUsage() {
     return usage;
+  }
+
+  public EnumerationItemDTO getEnumerationItemDTO() {
+    return enumerationItemDTO;
   }
 }
